#20.7000/bugfix/kir/dirty/6 Поправили типы в DirtyCheking
