+++ UI/Executor.ts
@@ -6,7 +6,7 @@ import RawMarkupNode from './_executor/_Expressions/RawMarkupNode';
 
 import * as _IGenerator from './_executor/_Markup/IGenerator';
 import * as _IGeneratorType from './_executor/_Markup/IGeneratorType';
-export { ITemplateNode } from './_executor/_Markup/IGeneratorType';
+export { ITemplateNode, TGeneratorNode } from './_executor/_Markup/IGeneratorType';
 import * as _IBuilder from './_executor/_Markup/IBuilder';
 
 // TODO: для работы executorCompatible
@@ -15,8 +15,6 @@ import * as _ForExecutorCompatible from './_executor/ForExecutorCompatible';
 /**
  * @author Тэн В.А.
  */
-
-export * from './_executor/_Markup/Vdom/IVdomType';
 
 export {
    htmlNode,
